+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # Origins of Life Changelog
+## 26.2.1.0
+- Adds breeding
+  - Each creature gets a random "Food" tag that it uses to breed. These are pulled from all vanilla entity's food tags.
+  - They are attracted to this item, like vanilla mobs
+  - When two creatures breed, the offspring is a combination of the two parents
+    - It selects a random model
+    - It selects random behaviors from the parents
+    - It selects random colors (or averages them, see configs below)
+    - It averages attribute values +/- a random percentage
+    - It selects random animation data between the parents, with small changes
+- Adds 2 new config values:
+  - `average_body_color` (false): Average parent's colors when breeding instead of picking one
+  - `average_eye_color` (false): Average parent's eye colors when breeding instead of picking one
+- Fossil blocks can now be put inside sulfur cubes
+  - Causes the same behavior has their non-fossil version
+
 ## 26.2.0.6-beta
 - Changed cage stack size to 1
 - Fixes an issue where display cases would get locked to displaying one entity